Removed shit code and distributed files into directories
