+++ PetFinder/Modules/CustomTabBar/CustomTabBarControllerViewController.swift
@@ -34,7 +34,7 @@ class CustomTabBarControllerViewController: UITabBarController {
 
     
     func setupTabBar() {
-        let profileView = ProfileSwiftUIView()
+        let profileView = ProfileView()
         let hostingController = UIHostingController(rootView: profileView)
         let mainView = MainView()
         let mainHostingController = UIHostingController(rootView: mainView)
@@ -42,7 +42,7 @@ class CustomTabBarControllerViewController: UITabBarController {
 
         
         let mainViewController = setupVC(viewController: mainHostingController, title: "Объявления", image:  setupImage(named: PFAssets.clipboard.image))
-        let createViewController = setupVC(viewController: ServicesViewController(), title: "Сервисы", image: setupImage(named: PFAssets.services.image))
+//        let createViewController = setupVC(viewController: ServicesViewController(), title: "Сервисы", image: setupImage(named: PFAssets.services.image))
         let nocreateViewController = setupVC(viewController: UIViewController(), title: "", image: nil)
         let profileViewController = setupVC(viewController: UIViewController(), title: "Избранное", image: setupImage(named: PFAssets.heart.image))
         let createViewControllerS = setupVC(viewController: hostingController, title: "Профиль", image: setupImage(named: PFAssets.profile.image))
