+++ PetFinder/Modules/MainScreen/AdsView/AdsView.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 import WrappingHStack
 
-struct AdNewView: View {
+struct AdsView: View {
     private var massivePet: [LostPetCard] = [
         LostPetCard(petName: "Джек", petType: .dog, adType: .lost, createPost: "Сегодня, 12:33", breed: "Хаски", color: "Серый", age: 2, gender: .male, lastSeenLocation: (55.7558, 37.6176), contactNumber: "+7 (123) 456-7890", additionalInfo: "Носит синий ошейник.", vaccinated: true, sterilised: true),
         LostPetCard(petName: "Мурка", petType: .cat, adType: .found, createPost: "Вчера, 10:13", breed: "Британская", color: "Черный", age: 5, gender: .female, lastSeenLocation: (59.9343, 30.3351), contactNumber: "+7 (987) 654-3210", additionalInfo: "В один обычный день, когда солнце ярко светило на улицах, произошло несчастье - собака по имени Бэйли исчезла. Это была веселая и игривая собака, которая всегда привносила радость и счастье в жизнь своих хозяев. Но сейчас все было иначе.Хозяева, Джон и Мэри, были в отчаянии. Они обошли все уголки своего района, обратились в местные приюты и объявили о пропаже в социальных сетях. Но нигде не было никаких следов Бэйли. Они были уверены, что собака была похищена, и их сердца были полны горя", vaccinated: false, sterilised: true),
@@ -36,5 +36,5 @@ struct AdNewView: View {
 }
 
 #Preview {
-    AdNewView()
+    AdsView()
 }
