+++ PetFinder/Modules/MainScreen/MainView.swift
@@ -14,7 +14,7 @@ struct MainView: View {
                 .background(Color(PFAssets.beige.color))
             AnimalCollectinView(massiveCell: [["sortImage", "Сортировка"], ["dog","Собаки"], ["cat","Кошки"], ["bird","Птицы"], ["mouse","Грызуны"], ["others","Прочее"]])
                 .padding(.leading, 16)
-            AdNewView()
+            AdsView()
                 .padding(.top, 24)
                 .padding(.leading, 16)
                 .padding(.trailing, 1)
