--- PetFinder/Modules/Services/ServicesTableView/Cell/ListServicesCollectionViewCell.swift
@@ -1,17 +1,0 @@
-//
-//  ListServicesCollectionViewCell.swift
-//  PetFinder
-//
-//  Created by Алексей Шамрей on 7.11.23.
-//
-
-import UIKit
-
-class ListServicesCollectionViewCell: UICollectionViewCell {
-    
-    func createUI() {
-        self.backgroundColor = UIColor(hex: 0x989898, alpha: 1)
-        self.layer.cornerRadius = 16
-    }
-
-}
